+++ kdrifting/models/__init__.py
@@ -2,7 +2,7 @@
 
 from kdrifting.models.convnext import ConvNextBase, ConvNextTiny, ConvNextV2
 from kdrifting.models.generator import DitGen, LightningDiT, build_generator_from_config
-from kdrifting.models.mae import MAEResNet, build_activation_function
+from kdrifting.models.mae import MAEResNet, build_activation_function, mae_from_metadata
 
 __all__ = [
     "ConvNextBase",
@@ -13,4 +13,5 @@ __all__ = [
     "MAEResNet",
     "build_activation_function",
     "build_generator_from_config",
+    "mae_from_metadata",
 ]